Move loaders to Loaders.js
Moved all the loaders to Loaders.js. Loaders are moved as is and have not been optimized. Loaders still contain same code in multiple locations but the reason for move is to make the Routes.js more readable
